+++ test/libsolidity/SemanticTest.cpp
@@ -68,10 +68,20 @@ namespace
 				// be signed. If an unsigned was detected in the expectations,
 				// but the actual result returned a signed, it would be formatted
 				// incorrectly.
-				if (*byteRange.begin() & 0x80)
-					resultStream << u2s(fromBigEndian<u256>(byteRange));
+				if (param.abiType.align == ABIType::AlignRight)
+				{
+					if (*byteRange.begin() & 0x80)
+						resultStream << u2s(fromBigEndian<u256>(byteRange));
+					else
+						resultStream << fromBigEndian<u256>(byteRange);
+				}
 				else
-					resultStream << fromBigEndian<u256>(byteRange);
+				{
+					byteRange.erase(
+						std::remove(byteRange.begin(), byteRange.end(), 0), byteRange.end()
+					);
+					resultStream << toHex(byteRange, HexPrefix::Add);
+				}
 				break;
 			case ABIType::Failure:
 				break;