+++ test/libsolidity/util/TestFileParser.h
@@ -56,6 +56,7 @@ namespace test
 	/* Literals & identifier */    \
 	T(Comment, "#", 0)             \
 	T(Number, "number", 0)         \
+	T(HexNumber, "hex_number", 0)  \
 	T(Identifier, "identifier", 0) \
 	/* type keywords */            \
 	K(Ether, "ether", 0)           \
@@ -104,7 +105,13 @@ struct ABIType
 		Failure,
 		None
 	};
+	enum Align {
+		AlignLeft,
+		AlignRight
+	};
+
 	Type type = ABIType::None;
+	Align align = ABIType::AlignRight;
 	size_t size = 0;
 };
 
@@ -281,7 +288,8 @@ private:
 
 		std::string scanComment();
 		std::string scanIdentifierOrKeyword();
-		std::string scanNumber();
+		std::string scanDecimalNumber();
+		std::string scanHexNumber();
 
 	private:
 		using TokenDesc = std::pair<Token, std::string>;
@@ -348,12 +356,20 @@ private:
 	/// # A nice comment. #
 	std::string parseComment();
 
-	/// Parses the current number literal.
-	std::string parseNumber();
+	/// Parses the current decimal number literal.
+	std::string parseDecimalNumber();
 
-	/// Tries to convert \param _literal to `uint256` and throws if
-	/// conversion fails.
+	/// Parses the current hex number literal.
+	std::string parseHexNumber();
+
+	/// Tries to convert \param _literal to right-aligned, padded `u256`
+	/// representation of the decimal number literal.
+	/// Throws if conversion fails.
 	u256 convertNumber(std::string const& _literal);
+
+	/// Tries to convert \param _literal to left-aligned, padded `bytes`
+	/// representation of the hex literal. Throws if conversion fails.
+	bytes convertHexNumber(std::string const& _literal);
 
 	/// A scanner instance
 	Scanner m_scanner;